refactoring : best organization for scalability
